Move util files for RGB pixels in separate folder
