+++ test/pixels/utils/RGBPixel/Helpers.cpp
