+++ test/pixels/utils/RGBPixel/Helpers.hpp
